requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.6.15
cffi==1.17.1
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.4.4
pip==24.3.1
propcache==0.3.2
pycparser==2.22
pydantic_core==2.33.2
pydantic==2.11.7
PyNaCl==1.5.0
typing_extensions==4.14.0
typing-inspection==0.4.1
yarl==1.20.1
